move component to bbui
