+++ packages/bbui/src/Input/CopyInput.svelte
