+++ packages/builder/src/components/settings/APIKeyModal.svelte
@@ -3,7 +3,7 @@
   import { Body, notifications } from "@budibase/bbui"
   import { auth } from "stores/portal"
   import { onMount } from "svelte"
-  import CopyInput from "components/common/inputs/CopyInput.svelte"
+  import CopyInput from "@budibase/bbui/src/Input/CopyInput.svelte"
 
   let apiKey = null
